// frontend/src/app/api/models/tour-log-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TourLogResponse {
  comment?: (string | null);
  difficulty?: (number | string);
  id?: (number | string);
  logDateTime?: string;
  rating?: (number | string);
  temperature?: (number | string | null);
  totalDistance?: (number | string);
  totalTime?: (number | string);
  tourId?: (number | string);
  weatherCondition?: (string | null);
}