// frontend/src/app/api/models/create-tour-log-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateTourLogRequest {
  comment?: (string | null);
  difficulty?: (number | string);
  logDateTime?: string;
  rating?: (number | string);
  temperature?: (number | string | null);
  totalDistance?: (number | string);
  totalTime?: (number | string);
  tourId?: (number | string);
  weatherCondition?: (string | null);
}